requirements.txt
apache-airflow==2.6.0 
pendulum==2.1.2 
kafka-python==2.0.2 
cassandra-driver==3.28.0 
pyspark==3.4.1
requests==2.31.0